requirement.txt
django==3.1.2
psycopg2-binary==2.8.6
python-dotenv==0.14.0
djangorestframework==3.12.1
django-cors-headers==3.5.0
PyJWT==1.71
pillow==8.0.1